// explore/explore-ui/src/app/modules/optimization-settings/objectives-settings/constants/objectives-settings-form-item-types.ts
export enum ObjectivesFormItemType {
    ENABLED = 'enabled',
    OBJECTIVE = 'objective',
    ENABLE_STRESS_SCENARIO = 'enableStressScenarios',
    ENABLE_ALPHA_SCORE = 'enableAlphaScore',
    DISABLED_OBJECTIVES = 'disabledObjectives',
    ENABLE_STRESS_SCENARIO_FOR_CUSTOM_DATE_RANGE = 'enableStressScenarioCustomDateRange'
}